Done with UI and stlying. But can do more stuff

+++ TodoList/View/ListView.swift
@@ -12,29 +12,38 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item)
+        ZStack {
+            if listViewModel.items.isEmpty {
+                NoItemView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                         }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
                     }
-                }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
+                    .listStyle(.plain)
             }
+        }
             .navigationTitle("Todo List 📝")
-            .listStyle(.plain)
             .toolbar {
                 ToolbarItem(placement: .topBarLeading) {
                     EditButton()
+                        .fontWeight(.semibold)
                 }
                 
                 ToolbarItem(placement: .topBarTrailing) {
                     NavigationLink("Add") {
                         AddView()
                     }
+                    .fontWeight(.semibold)
                 }
             }
         }
